Originally I had:
```
(
  (_
    [
      ":"
      "="
    ] @_.leading
    value: (_) @value
  ) @_.domain
  (#not-type? @_.domain subscript)
)
```

to handle all cases of `value` in arguments or dictionaries:

```
def func(b=True, c=True):
    pass
d1 = {"a": 1, "b": 2, "c": 3}
_ = {value: key for (key, value) in d1.items()}
```

but i wanted to change as there was conflict between ":" and "=". Indeed, `chunk value` here would remove ` str = "hello"` instead of `= "hello"`

```
def my_funk(value: str = "hello", other: bool=False) -> str:
    print(value)
```

So I wanted to split it into the following:

```
(parameters
  (_
    "=" @_.leading
    value: (_) @value
  ) @_.domain
)
```

for

```
def func(b=True, c=True):
```

and

```
(dictionary
  (_
    ":" @_.leading
    value: (_) @value
  ) @_.domain
)
```

for

```
d1 = {"a": 1, "b": 2, "c": 3}
```

and

```
(dictionary_comprehension
  (_
    ":" @_.leading
    value: (_) @value
  ) @_.domain
)
```

for

```
_ = {value: key for (key, value) in d1.items()}
```

And so it seems annoying to have the repedition for `dictionary` and `dictionary_comprehension` as we can't define something like:

```
([dictionary dictionary_comprehension]
  (_
    ":" @_.leading
    value: (_) @value
  ) @_.domain
)
```

as we get a syntax error, as is not allowed by scm.

And so I ended up defining a new predicate `#has-type` and I was able to do this:

```
(
  (_
    (pair
      ":" @_.leading
      value: (_) @value
    ) @_.domain
  ) @_.tmp
  (#has-type? @_.tmp dictionary dictionary_comprehension)
)
```

and it worked:

```
d1 = {"a": 1, "b": 2, "c": 3}
_ = {value: key for (key, value) in d1.items()}
```

NOTE: at the end, there was an easier way to define the value scope for `dictionary` and `dictionary_comprehension` which is:

```
(pair
  ":" @_.leading
  value: (_) @value
) @_.domain
```

so we didn't need `#has-type` but it could become handy another day.